Fix indexing issues due to copy field
Fixed the multiValued parameter in the api request to convert binary to
string accurately. Added code to skip adding fields to copy based
on the include_in_all mapping value. Also added a function parameter for
the stored api request parameter since the copy field does not need to
be stored.